Removed "int" wrapper routines for brun and evnt so that run number and event number are forced to be int32_t and uint64_t respectively. The previous JANA version had both forms as an option to ease the transition. This completes the transition by requiring code to adhere to new format.
git-svn-id: https://phys12svn.jlab.org/repos/JANA@2376 c5ed4466-e916-0410-8347-b3263e9c103d

+++ src/JANA/JEventSink.h
@@ -31,8 +31,8 @@ class JEventSink:public JEventProcessor{
 	protected:
 		virtual jerror_t init(void){return NOERROR;}				///< Called once at program start.
 		jerror_t brun(JEventLoop *eventLoop, int runnumber);	///< Called everytime a new run number is detected.
-		virtual jerror_t brun_sink(JEventLoop *loop, int runnumber)=0;
-		virtual jerror_t evnt(JEventLoop *loop, int eventnumber){return NOERROR;}	///< Called every event.
+		virtual jerror_t brun_sink(JEventLoop *loop, int32_t runnumber)=0;
+		virtual jerror_t evnt(JEventLoop *loop, uint64_t eventnumber){return NOERROR;}	///< Called every event.
 		virtual jerror_t erun(void){return NOERROR;}				///< Called everytime run number changes, provided brun has been called.
 		virtual jerror_t fini(void){return NOERROR;}				///< Called after last event of last event source has been processed.
 		void AddToWriteList(string name, string tag);
