+++ src/unit_tests/associated_objects/JEventProcessor_TestProcessor.h
@@ -93,11 +93,11 @@ _DBG_<<"---- "<<objs[0]->className()<<" size="<<objs.size()<<endl;
 
 	private:
 		jerror_t init(void){return NOERROR;}
-		jerror_t brun(jana::JEventLoop *eventLoop, int runnumber){return NOERROR;}
+		jerror_t brun(jana::JEventLoop *eventLoop, int32_t runnumber){return NOERROR;}
 		jerror_t erun(void){return NOERROR;}
 		jerror_t fini(void){return NOERROR;}
 
-		jerror_t evnt(jana::JEventLoop *loop, int eventnumber){
+		jerror_t evnt(jana::JEventLoop *loop, uint64_t eventnumber){
 			
 			// Get objects
 			loop->Get(objsA);
