+++ src/unit_tests/associated_objects/TestClassA_factory.h
@@ -19,11 +19,11 @@ class TestClassA_factory:public jana::JFactory<TestClassA>{
 
 	private:
 		jerror_t init(void){return NOERROR;}
-		jerror_t brun(jana::JEventLoop *eventLoop, int runnumber){return NOERROR;}
+		jerror_t brun(jana::JEventLoop *eventLoop, int32_t runnumber){return NOERROR;}
 		jerror_t erun(void){return NOERROR;}
 		jerror_t fini(void){return NOERROR;}
 
-		jerror_t evnt(jana::JEventLoop *eventLoop, int eventnumber){
+		jerror_t evnt(jana::JEventLoop *eventLoop, uint64_t eventnumber){
 
 			// Just create 8 objects of type TestClassA
 			for(int i=0; i<8; i++){
